extracted out decliner input into its own component
